// src/store/article/article-action-types.ts
export enum ArticleActionTypes {
  ARTICLE_RESET_STATE = 'ARTICLE_RESET_STATE',
  ARTICLE_EDIT_REMOVE_TAG = 'ARTICLE_EDIT_REMOVE_TAG',
  ARTICLE_EDIT_ADD_TAG = 'ARTICLE_EDIT_ADD_TAG',
  ARTICLE_EDIT = 'ARTICLE_EDIT',
  ARTICLE_DELETE = 'ARTICLE_DELETE',
  ARTICLE_PUBLISH = 'ARTICLE_PUBLISH',
  FAVORITE_REMOVE = 'FAVORITE_REMOVE',
  FAVORITE_ADD = 'FAVORITE_ADD',
  COMMENT_DESTROY = 'COMMENT_DESTROY',
  COMMENT_CREATE = 'COMMENT_CREATE',
  FETCH_COMMENTS = 'FETCH_COMMENTS',
  FETCH_ARTICLE = 'FETCH_ARTICLE'
}